+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.3
+
+- Adds ability to choose if top, bottom, both sides, or neither side of the sheet should have a border radius.
+- Fixes default opacity.
+
 ## 0.0.2
 
 - README updates.